+++ test/sefaz/test_nfe.rb
@@ -12,28 +12,32 @@ class TestNFE < Minitest::Test
   end
 
   def test_if_the_statusDoServico_is_working
-    response = @webService.statusDoServico
-    assert_equal response[:nfe_result_msg][:ret_cons_stat_serv][:c_stat], "107"
+    xml, hash = @webService.statusDoServico
+    refute_nil xml
+    assert_equal hash[:nfeResultMsg][:retConsStatServ][:cStat], "107"
   end
 
   def test_is_the_consultarNF_is_working
     chaveNF = "35221221684155000164550010000002361125429343"
-    response = @webService.consultarNF(chaveNF)
-    assert_equal response[:nfe_result_msg][:ret_cons_sit_n_fe][:c_stat], "101"
+    xml, hash = @webService.consultarNF(chaveNF)
+    refute_nil xml
+    assert_equal hash[:nfeResultMsg][:retConsSitNFe][:cStat], "101"
   end
 
   def test_is_the_consultarCadastro_is_working
     nroDocumento = "01.014.625/0001-75"
     tpDocumento = "CNPJ"
     uf = "SP"
-    response = @webService.consultarCadastro(nroDocumento, tpDocumento, uf)
-    assert_equal response[:nfe_result_msg][:ret_cons_cad][:inf_cons][:c_stat], "259"
+    xml, hash = @webService.consultarCadastro(nroDocumento, tpDocumento, uf)
+    refute_nil xml
+    assert_equal hash[:nfeResultMsg][:retConsCad][:infCons][:cStat], "259"
   end
 
   def test_is_the_consultarRecibo_is_working
     numRecibo = "351000171600547"
-    response = @webService.consultarRecibo(numRecibo)
-    assert_equal response[:nfe_result_msg][:ret_cons_reci_n_fe][:c_stat], "104"
+    xml, hash = @webService.consultarRecibo(numRecibo)
+    refute_nil xml
+    assert_equal hash[:nfeResultMsg][:retConsReciNFe][:cStat], "104"
   end
 
 end